Don't run LVM cleanups until after taking the lock
Previously, this could try to deactivate a VG it hadn't created yet,
usually while some other running pcib held the lock and was using a VG
by that name. This would cause the build to fail, and other cleanups
not to run.

By not registering the cleanup until after we've taken the lock, we
ensure that we're in control of the VG before we go cleaning it up.