Always use nullglob
The lack of nullglob breaks plugins without tasks (e.g., lib/ruby), as
the glob in register_plugin_tasks expands to itself. This is likely to
be useful in other cases as well, so just set it everywhere.

Also quote arguments to unset, since [ and ] are glob characters and
thus "unset foo[$x]" will have its argument expanded to nothing. This
is a good idea regardless, just in case the current directory happens
to contain a file matching "foo[$x]", but I hadn't noticed it until
now.