os/debian: Perform a minimal debootstrap
This is more consistent with the minimal approach taken by os/centos.
It enables other plugins to be composable on top of this one by
installing additional packages, rather than getting a monolithic base
install that may or may not be suitable for every use case.

Since --variant=minbase doesn't pull in dhclient, install that whenever
a DHCP interface is requested.