Enable multiple LVM builds to run in parallel
Previously, we put a lock around most of the build, from VG creation to
VG disablement. This made parallelism possible, but mostly pointless.

Instead, use a randomly-generated VG name for most of the build, and
only serialise a single task which renames and then deactivates the VG
at the end.