Sleep after making any changes to device-mapper
This is a bit of a weird issue, where occasionally an image build will
fail because a change that was just made hasn't shown up in /dev yet. I
plan to revisit this and see if I can work out a proper fix at some
point, but for now, this allows image builds to run more reliably.