os/centos: Ensure /var/run exists before bootstrap
Due to a combination of factors, /var/run does not always get created
as a symlink during yum bootstrap. The "filesystem" package is supposed
to create it, but will only do so if /var already exists, which won't
be the case if (as on Debian) rpm's database isn't in /var by default.

The issue of rpm's misplaced database will be addressed in a following
commit, but it's worthwhile also making sure we have /var/run before
installing any packages, just in case.