os/centos: Ensure the rpm db is in the right place
Some distributions, such as Debian, patch rpm to use a different
default database file. This results in the bootstrapped system not
having an rpm database where it's supposed to be, which had been
previously worked around by installing packages again inside the
chroot.

Instead, we will now fail hard if rpm is misconfigured in this way.
It's up to the local sysadmin to configure rpm's dbpath however best
suits their installation.