os/centos: Provide a hook to add bootstrap repos
It may be desirable for third-party plugins to add package repositories
early in the build; for example, to replace packages in the base OS
with customised versions.

To accomplish this, we move the bootstrap repo generation logic to
priority 18, and provide a function that can be called at priority 19
in time for the yum bootstrap at priority 20. There should be no
functional change to the generated image if no third-party module uses
this hook.