os/centos: Create /etc/sysconfig/kernel
/etc/sysconfig/kernel belongs to, but is not automatically created by,
the "grubby" package. It is normally created by Anaconda during
installation. Leaving this file missing means that the GRUB
configuration will not be updated when new kernels are installed, so
the system will continue to boot with its original kernel without
manual intervention.

Here, we just blat out a simple /etc/sysconfig/kernel that configures
grubby to update GRUB for new kernels.